moved grid script into namespace and folder
